Add files via upload
this is a program where the entered string is checked if it is a pangram or not. here we compare the entered string with another string called letters which contains all the alphabets. so each letter of the entered string is compared with the letters string. if all alphabet are present, then the entered string is a pangram.